+++ src/judgeval/common/tracer/__init__.py
@@ -1,21 +1,26 @@
-from judgeval.common.tracer.tracer import (
-    BackgroundSpanService,
+from judgeval.common.tracer.trace_manager import TraceManagerClient
+from judgeval.common.tracer.background_span import BackgroundSpanService
+from judgeval.common.tracer.core import (
     TraceClient,
-    TraceManagerClient,
-    TraceThreadPoolExecutor,
     Tracer,
     wrap,
     current_span_var,
     current_trace_var,
+    TraceSpan,
+    SpanType,
+    cost_per_token,
 )
+
 
 __all__ = [
     "Tracer",
     "TraceClient",
     "TraceManagerClient",
     "BackgroundSpanService",
-    "TraceThreadPoolExecutor",
     "wrap",
     "current_span_var",
     "current_trace_var",
+    "TraceSpan",
+    "SpanType",
+    "cost_per_token",
 ]
