+++ src/judgeval/common/tracer/core.py
@@ -3,6 +3,7 @@ Tracing system for judgeval that allows for function tracing using decorators.
 """
 
 from __future__ import annotations
+
 import asyncio
 import functools
 import inspect
@@ -14,14 +15,12 @@ import uuid
 import warnings
 import contextvars
 import sys
-import json
 from contextlib import (
     contextmanager,
     AbstractAsyncContextManager,
     AbstractContextManager,
 )
 from datetime import datetime, timezone
-from http import HTTPStatus
 from typing import (
     Any,
     Callable,
@@ -35,25 +34,17 @@ from typing import (
     AsyncGenerator,
     TypeAlias,
 )
-from rich import print as rprint
 import types
 
-from requests import RequestException
+from judgeval.common.tracer import TraceManagerClient
+from judgeval.common.tracer.background_span import BackgroundSpanService
 from judgeval.common.tracer.constants import TRACE_FILEPATH_BLOCKLIST
-from judgeval.utils.requests import requests
 from litellm import cost_per_token as _original_cost_per_token
 from openai import OpenAI, AsyncOpenAI
 from together import Together, AsyncTogether
 from anthropic import Anthropic, AsyncAnthropic
 from google import genai
 
-from judgeval.constants import (
-    JUDGMENT_TRACES_SAVE_API_URL,
-    JUDGMENT_TRACES_UPSERT_API_URL,
-    JUDGMENT_TRACES_FETCH_API_URL,
-    JUDGMENT_TRACES_SPANS_BATCH_API_URL,
-    JUDGMENT_TRACES_EVALUATION_RUNS_BATCH_API_URL,
-)
 from judgeval.data import Example, Trace, TraceSpan, TraceUsage
 from judgeval.scorers import APIJudgmentScorer, JudgevalScorer
 from judgeval.rules import Rule
@@ -62,8 +53,6 @@ from judgeval.common.utils import ExcInfo, validate_api_key
 from judgeval.common.exceptions import JudgmentAPIError
 
 from collections.abc import Iterator, AsyncIterator
-import queue
-import atexit
 
 current_trace_var = contextvars.ContextVar[Optional["TraceClient"]](
     "current_trace", default=None
@@ -81,194 +70,6 @@ ApiClient: TypeAlias = Union[
     genai.client.AsyncClient,
 ]
 SpanType = Union[Literal["span", "tool", "function", "llm", "evaluation", "chain"], str]
-
-
-class TraceManagerClient:
-    """
-    Client for handling trace endpoints with the Judgment API
-
-
-    Operations include:
-    - Fetching a trace by id
-    - Saving a trace
-    - Deleting a trace
-    """
-
-    def __init__(
-        self,
-        judgment_api_key: str,
-        organization_id: str,
-        tracer: Optional["Tracer"] = None,
-    ):
-        self.judgment_api_key = judgment_api_key
-        self.organization_id = organization_id
-        self.tracer = tracer
-
-    def fetch_trace(self, trace_id: str):
-        """
-        Fetch a trace by its id
-        """
-        response = requests.post(
-            JUDGMENT_TRACES_FETCH_API_URL,
-            json={
-                "trace_id": trace_id,
-            },
-            headers={
-                "Content-Type": "application/json",
-                "Authorization": f"Bearer {self.judgment_api_key}",
-                "X-Organization-Id": self.organization_id,
-            },
-            verify=True,
-        )
-
-        if response.status_code != HTTPStatus.OK:
-            raise ValueError(f"Failed to fetch traces: {response.text}")
-
-        return response.json()
-
-    def save_trace(
-        self, trace_data: dict, offline_mode: bool = False, final_save: bool = True
-    ):
-        """
-        Saves a trace to the Judgment Supabase and optionally to S3 if configured.
-
-        Args:
-            trace_data: The trace data to save
-            offline_mode: Whether running in offline mode
-            final_save: Whether this is the final save (controls S3 saving)
-            NOTE we save empty traces in order to properly handle async operations; we need something in the DB to associate the async results with
-
-        Returns:
-            dict: Server response containing UI URL and other metadata
-        """
-        # Save to Judgment API
-
-        def fallback_encoder(obj):
-            """
-            Custom JSON encoder fallback.
-            Tries to use obj.__repr__(), then str(obj) if that fails or for a simpler string.
-            You can choose which one you prefer or try them in sequence.
-            """
-            try:
-                # Option 1: Prefer __repr__ for a more detailed representation
-                return repr(obj)
-            except Exception:
-                # Option 2: Fallback to str() if __repr__ fails or if you prefer str()
-                try:
-                    return str(obj)
-                except Exception as e:
-                    # If both fail, you might return a placeholder or re-raise
-                    return f"<Unserializable object of type {type(obj).__name__}: {e}>"
-
-        serialized_trace_data = json.dumps(trace_data, default=fallback_encoder)
-        response = requests.post(
-            JUDGMENT_TRACES_SAVE_API_URL,
-            data=serialized_trace_data,
-            headers={
-                "Content-Type": "application/json",
-                "Authorization": f"Bearer {self.judgment_api_key}",
-                "X-Organization-Id": self.organization_id,
-            },
-            verify=True,
-        )
-
-        if response.status_code == HTTPStatus.BAD_REQUEST:
-            raise ValueError(
-                f"Failed to save trace data: Check your Trace name for conflicts, set overwrite=True to overwrite existing traces: {response.text}"
-            )
-        elif response.status_code != HTTPStatus.OK:
-            raise ValueError(f"Failed to save trace data: {response.text}")
-
-        # Parse server response
-        server_response = response.json()
-
-        # If S3 storage is enabled, save to S3 only on final save
-        if self.tracer and self.tracer.use_s3 and final_save:
-            try:
-                s3_key = self.tracer.s3_storage.save_trace(
-                    trace_data=trace_data,
-                    trace_id=trace_data["trace_id"],
-                    project_name=trace_data["project_name"],
-                )
-                print(f"Trace also saved to S3 at key: {s3_key}")
-            except Exception as e:
-                warnings.warn(f"Failed to save trace to S3: {str(e)}")
-
-        if not offline_mode and "ui_results_url" in server_response:
-            pretty_str = f"\n🔍 You can view your trace data here: [rgb(106,0,255)][link={server_response['ui_results_url']}]View Trace[/link]\n"
-            rprint(pretty_str)
-
-        return server_response
-
-    def upsert_trace(
-        self,
-        trace_data: dict,
-        offline_mode: bool = False,
-        show_link: bool = True,
-        final_save: bool = True,
-    ):
-        """
-        Upserts a trace to the Judgment API (always overwrites if exists).
-
-        Args:
-            trace_data: The trace data to upsert
-            offline_mode: Whether running in offline mode
-            show_link: Whether to show the UI link (for live tracing)
-            final_save: Whether this is the final save (controls S3 saving)
-
-        Returns:
-            dict: Server response containing UI URL and other metadata
-        """
-
-        def fallback_encoder(obj):
-            """
-            Custom JSON encoder fallback.
-            Tries to use obj.__repr__(), then str(obj) if that fails or for a simpler string.
-            """
-            try:
-                return repr(obj)
-            except Exception:
-                try:
-                    return str(obj)
-                except Exception as e:
-                    return f"<Unserializable object of type {type(obj).__name__}: {e}>"
-
-        serialized_trace_data = json.dumps(trace_data, default=fallback_encoder)
-
-        response = requests.post(
-            JUDGMENT_TRACES_UPSERT_API_URL,
-            data=serialized_trace_data,
-            headers={
-                "Content-Type": "application/json",
-                "Authorization": f"Bearer {self.judgment_api_key}",
-                "X-Organization-Id": self.organization_id,
-            },
-            verify=True,
-        )
-
-        if response.status_code != HTTPStatus.OK:
-            raise ValueError(f"Failed to upsert trace data: {response.text}")
-
-        # Parse server response
-        server_response = response.json()
-
-        # If S3 storage is enabled, save to S3 only on final save
-        if self.tracer and self.tracer.use_s3 and final_save:
-            try:
-                s3_key = self.tracer.s3_storage.save_trace(
-                    trace_data=trace_data,
-                    trace_id=trace_data["trace_id"],
-                    project_name=trace_data["project_name"],
-                )
-                print(f"Trace also saved to S3 at key: {s3_key}")
-            except Exception as e:
-                warnings.warn(f"Failed to save trace to S3: {str(e)}")
-
-        if not offline_mode and show_link and "ui_results_url" in server_response:
-            pretty_str = f"\n🔍 You can view your trace data here: [rgb(106,0,255)][link={server_response['ui_results_url']}]View Trace[/link]\n"
-            rprint(pretty_str)
-
-        return server_response
 
 
 class TraceClient:
@@ -297,21 +98,19 @@ class TraceClient:
         self.enable_evaluations = enable_evaluations
         self.parent_trace_id = parent_trace_id
         self.parent_name = parent_name
-        self.customer_id: Optional[str] = None  # Added customer_id attribute
-        self.tags: Optional[List[str]] = None  # Added tags attribute
-        self.has_notification: Optional[bool] = False  # Initialize has_notification
+        self.customer_id: Optional[str] = None
+        self.tags: Optional[List[str]] = None
+        self.has_notification: Optional[bool] = False
         self.trace_spans: List[TraceSpan] = []
         self.span_id_to_span: Dict[str, TraceSpan] = {}
         self.evaluation_runs: List[EvaluationRun] = []
-        self.start_time: Optional[float] = (
-            None  # Will be set after first successful save
-        )
+        self.start_time: Optional[float] = None
+
         self.trace_manager_client = TraceManagerClient(
             tracer.api_key, tracer.organization_id, tracer
         )
-        self._span_depths: Dict[str, int] = {}  # NEW: To track depth of active spans
+        self._span_depths: Dict[str, int] = {}
 
-        # Get background span service from tracer
         self.background_span_service = (
             tracer.get_background_span_service() if tracer else None
         )
@@ -463,7 +262,6 @@ class TraceClient:
 
         span_id_to_use = span_id if span_id is not None else self.get_current_span()
 
-        # Combine the trace-level rules with any evaluation-specific rules)
         eval_run = EvaluationRun(
             organization_id=self.tracer.organization_id,
             project_name=self.project_name,
@@ -478,7 +276,7 @@ class TraceClient:
             trace_span_id=span_id_to_use,
         )
 
-        self.add_eval_run(eval_run, start_time)  # Pass start_time to record_evaluation
+        self.add_eval_run(eval_run, start_time)
 
         # Queue evaluation run through background service
         if self.background_span_service and span_id_to_use:
@@ -490,10 +288,7 @@ class TraceClient:
                 )
 
     def add_eval_run(self, eval_run: EvaluationRun, start_time: float):
-        # --- Modification: Use span_id from eval_run ---
-        current_span_id = eval_run.trace_span_id  # Get ID from the eval_run object
-        # print(f"[TraceClient.add_eval_run] Using span_id from eval_run: {current_span_id}")
-        # --- End Modification ---
+        current_span_id = eval_run.trace_span_id
 
         if current_span_id:
             span = self.span_id_to_span[current_span_id]
@@ -832,337 +627,6 @@ def _capture_exception_for_trace(
             )
 
 
-class BackgroundSpanService:
-    """
-    Background service for queueing and batching trace spans for efficient saving.
-
-    This service:
-    - Queues spans as they complete
-    - Batches them for efficient network usage
-    - Sends spans periodically or when batches reach a certain size
-    - Handles automatic flushing when the main event terminates
-    """
-
-    def __init__(
-        self,
-        judgment_api_key: str,
-        organization_id: str,
-        batch_size: int = 10,
-        flush_interval: float = 5.0,
-        num_workers: int = 1,
-    ):
-        """
-        Initialize the background span service.
-
-        Args:
-            judgment_api_key: API key for Judgment service
-            organization_id: Organization ID
-            batch_size: Number of spans to batch before sending (default: 10)
-            flush_interval: Time in seconds between automatic flushes (default: 5.0)
-            num_workers: Number of worker threads to process the queue (default: 1)
-        """
-        self.judgment_api_key = judgment_api_key
-        self.organization_id = organization_id
-        self.batch_size = batch_size
-        self.flush_interval = flush_interval
-        self.num_workers = max(1, num_workers)
-
-        # Queue for pending spans
-        self._span_queue: queue.Queue[Dict[str, Any]] = queue.Queue()
-
-        # Background threads for processing spans
-        self._worker_threads: List[threading.Thread] = []
-        self._shutdown_event = threading.Event()
-
-        # Track spans that have been sent
-        # self._sent_spans = set()
-
-        # Register cleanup on exit
-        atexit.register(self.shutdown)
-
-        # Start the background workers
-        self._start_workers()
-
-    def _start_workers(self):
-        """Start the background worker threads."""
-        for i in range(self.num_workers):
-            if len(self._worker_threads) < self.num_workers:
-                worker_thread = threading.Thread(
-                    target=self._worker_loop, daemon=True, name=f"SpanWorker-{i + 1}"
-                )
-                worker_thread.start()
-                self._worker_threads.append(worker_thread)
-
-    def _worker_loop(self):
-        """Main worker loop that processes spans in batches."""
-        batch = []
-        last_flush_time = time.time()
-        # Track how many tasks we've taken from queue but not marked done
-        pending_task_count = 0
-
-        while not self._shutdown_event.is_set() or self._span_queue.qsize() > 0:
-            try:
-                # First, do a blocking get to wait for at least one item
-                if not batch:  # Only block if we don't have items already
-                    try:
-                        span_data = self._span_queue.get(timeout=1.0)
-                        batch.append(span_data)
-                        pending_task_count += 1
-                    except queue.Empty:
-                        # No new spans, continue to check for flush conditions
-                        pass
-
-                # Then, do non-blocking gets to drain any additional available items
-                # up to our batch size limit
-                while len(batch) < self.batch_size:
-                    try:
-                        span_data = self._span_queue.get_nowait()  # Non-blocking
-                        batch.append(span_data)
-                        pending_task_count += 1
-                    except queue.Empty:
-                        # No more items immediately available
-                        break
-
-                current_time = time.time()
-                should_flush = len(batch) >= self.batch_size or (
-                    batch and (current_time - last_flush_time) >= self.flush_interval
-                )
-
-                if should_flush and batch:
-                    self._send_batch(batch)
-
-                    # Only mark tasks as done after successful sending
-                    for _ in range(pending_task_count):
-                        self._span_queue.task_done()
-                    pending_task_count = 0  # Reset counter
-
-                    batch.clear()
-                    last_flush_time = current_time
-
-            except Exception as e:
-                warnings.warn(f"Error in span service worker loop: {e}")
-                # On error, still need to mark tasks as done to prevent hanging
-                for _ in range(pending_task_count):
-                    self._span_queue.task_done()
-                pending_task_count = 0
-                batch.clear()
-
-        # Final flush on shutdown
-        if batch:
-            self._send_batch(batch)
-            # Mark remaining tasks as done
-            for _ in range(pending_task_count):
-                self._span_queue.task_done()
-
-    def _send_batch(self, batch: List[Dict[str, Any]]):
-        """
-        Send a batch of spans to the server.
-
-        Args:
-            batch: List of span dictionaries to send
-        """
-        if not batch:
-            return
-
-        try:
-            # Group items by type for different endpoints
-            spans_to_send = []
-            evaluation_runs_to_send = []
-
-            for item in batch:
-                if item["type"] == "span":
-                    spans_to_send.append(item["data"])
-                elif item["type"] == "evaluation_run":
-                    evaluation_runs_to_send.append(item["data"])
-
-            # Send spans if any
-            if spans_to_send:
-                self._send_spans_batch(spans_to_send)
-
-            # Send evaluation runs if any
-            if evaluation_runs_to_send:
-                self._send_evaluation_runs_batch(evaluation_runs_to_send)
-
-        except Exception as e:
-            warnings.warn(f"Failed to send batch: {e}")
-
-    def _send_spans_batch(self, spans: List[Dict[str, Any]]):
-        """Send a batch of spans to the spans endpoint."""
-        payload = {"spans": spans, "organization_id": self.organization_id}
-
-        # Serialize with fallback encoder
-        def fallback_encoder(obj):
-            try:
-                return repr(obj)
-            except Exception:
-                try:
-                    return str(obj)
-                except Exception as e:
-                    return f"<Unserializable object of type {type(obj).__name__}: {e}>"
-
-        try:
-            serialized_data = json.dumps(payload, default=fallback_encoder)
-
-            # Send the actual HTTP request to the batch endpoint
-            response = requests.post(
-                JUDGMENT_TRACES_SPANS_BATCH_API_URL,
-                data=serialized_data,
-                headers={
-                    "Content-Type": "application/json",
-                    "Authorization": f"Bearer {self.judgment_api_key}",
-                    "X-Organization-Id": self.organization_id,
-                },
-                verify=True,
-                timeout=30,  # Add timeout to prevent hanging
-            )
-
-            if response.status_code != HTTPStatus.OK:
-                warnings.warn(
-                    f"Failed to send spans batch: HTTP {response.status_code} - {response.text}"
-                )
-
-        except RequestException as e:
-            warnings.warn(f"Network error sending spans batch: {e}")
-        except Exception as e:
-            warnings.warn(f"Failed to serialize or send spans batch: {e}")
-
-    def _send_evaluation_runs_batch(self, evaluation_runs: List[Dict[str, Any]]):
-        """Send a batch of evaluation runs with their associated span data to the endpoint."""
-        # Structure payload to include both evaluation run data and span data
-        evaluation_entries = []
-        for eval_data in evaluation_runs:
-            # eval_data already contains the evaluation run data (no need to access ['data'])
-            entry = {
-                "evaluation_run": {
-                    # Extract evaluation run fields (excluding span-specific fields)
-                    key: value
-                    for key, value in eval_data.items()
-                    if key not in ["associated_span_id", "span_data", "queued_at"]
-                },
-                "associated_span": {
-                    "span_id": eval_data.get("associated_span_id"),
-                    "span_data": eval_data.get("span_data"),
-                },
-                "queued_at": eval_data.get("queued_at"),
-            }
-            evaluation_entries.append(entry)
-
-        payload = {
-            "organization_id": self.organization_id,
-            "evaluation_entries": evaluation_entries,  # Each entry contains both eval run + span data
-        }
-
-        # Serialize with fallback encoder
-        def fallback_encoder(obj):
-            try:
-                return repr(obj)
-            except Exception:
-                try:
-                    return str(obj)
-                except Exception as e:
-                    return f"<Unserializable object of type {type(obj).__name__}: {e}>"
-
-        try:
-            serialized_data = json.dumps(payload, default=fallback_encoder)
-
-            # Send the actual HTTP request to the batch endpoint
-            response = requests.post(
-                JUDGMENT_TRACES_EVALUATION_RUNS_BATCH_API_URL,
-                data=serialized_data,
-                headers={
-                    "Content-Type": "application/json",
-                    "Authorization": f"Bearer {self.judgment_api_key}",
-                    "X-Organization-Id": self.organization_id,
-                },
-                verify=True,
-                timeout=30,
-            )
-
-            if response.status_code != HTTPStatus.OK:
-                warnings.warn(
-                    f"Failed to send evaluation runs batch: HTTP {response.status_code} - {response.text}"
-                )
-
-        except RequestException as e:
-            warnings.warn(f"Network error sending evaluation runs batch: {e}")
-        except Exception as e:
-            warnings.warn(f"Failed to send evaluation runs batch: {e}")
-
-    def queue_span(self, span: TraceSpan, span_state: str = "input"):
-        """
-        Queue a span for background sending.
-
-        Args:
-            span: The TraceSpan object to queue
-            span_state: State of the span ("input", "output", "completed")
-        """
-        if not self._shutdown_event.is_set():
-            span_data = {
-                "type": "span",
-                "data": {
-                    **span.model_dump(),
-                    "span_state": span_state,
-                    "queued_at": time.time(),
-                },
-            }
-            self._span_queue.put(span_data)
-
-    def queue_evaluation_run(
-        self, evaluation_run: EvaluationRun, span_id: str, span_data: TraceSpan
-    ):
-        """
-        Queue an evaluation run for background sending.
-
-        Args:
-            evaluation_run: The EvaluationRun object to queue
-            span_id: The span ID associated with this evaluation run
-            span_data: The span data at the time of evaluation (to avoid race conditions)
-        """
-        if not self._shutdown_event.is_set():
-            eval_data = {
-                "type": "evaluation_run",
-                "data": {
-                    **evaluation_run.model_dump(),
-                    "associated_span_id": span_id,
-                    "span_data": span_data.model_dump(),
-                    "queued_at": time.time(),
-                },
-            }
-            self._span_queue.put(eval_data)
-
-    def flush(self):
-        """Force immediate sending of all queued spans."""
-        try:
-            # Wait for the queue to be processed
-            self._span_queue.join()
-        except Exception as e:
-            warnings.warn(f"Error during flush: {e}")
-
-    def shutdown(self):
-        """Shutdown the background service and flush remaining spans."""
-        if self._shutdown_event.is_set():
-            return
-
-        try:
-            # Signal shutdown to stop new items from being queued
-            self._shutdown_event.set()
-
-            # Try to flush any remaining spans
-            try:
-                self.flush()
-            except Exception as e:
-                warnings.warn(f"Error during final flush: {e}")
-        except Exception as e:
-            warnings.warn(f"Error during BackgroundSpanService shutdown: {e}")
-        finally:
-            # Clear the worker threads list (daemon threads will be killed automatically)
-            self._worker_threads.clear()
-
-    def get_queue_size(self) -> int:
-        """Get the current size of the span queue."""
-        return self._span_queue.qsize()
-
-
 class _DeepTracer:
     _instance: Optional["_DeepTracer"] = None
     _lock: threading.Lock = threading.Lock()
@@ -1435,17 +899,13 @@ class Tracer:
     # TODO: Keep track of cross-context state for current trace and current span ID solely through class variables instead of instance variables?
     # Should be fine to do so as long as we keep Tracer as a singleton
     current_trace: Optional[TraceClient] = None
-    # current_span_id: Optional[str] = None
-
-    trace_across_async_contexts: bool = (
-        False  # BY default, we don't trace across async contexts
-    )
+    trace_across_async_contexts: bool = False
 
     def __init__(
         self,
         api_key: str | None = os.getenv("JUDGMENT_API_KEY"),
         project_name: str | None = None,
-        rules: Optional[List[Rule]] = None,  # Added rules parameter
+        rules: Optional[List[Rule]] = None,
         organization_id: str | None = os.getenv("JUDGMENT_ORG_ID"),
         enable_monitoring: bool = os.getenv("JUDGMENT_MONITORING", "true").lower()
         == "true",
@@ -1458,8 +918,8 @@ class Tracer:
         s3_aws_secret_access_key: Optional[str] = None,
         s3_region_name: Optional[str] = None,
         offline_mode: bool = False,
-        deep_tracing: bool = False,  # Deep tracing is disabled by default
-        trace_across_async_contexts: bool = False,  # BY default, we don't trace across async contexts
+        deep_tracing: bool = False,
+        trace_across_async_contexts: bool = False,
         # Background span service configuration
         enable_background_spans: bool = True,  # Enable background span service by default
         span_batch_size: int = 50,  # Number of spans to batch before sending
@@ -1481,13 +941,12 @@ class Tracer:
         self.api_key: str = api_key
         self.project_name: str = project_name or str(uuid.uuid4())
         self.organization_id: str = organization_id
-        self.rules: List[Rule] = rules or []  # Store rules at tracer level
+        self.rules: List[Rule] = rules or []
         self.traces: List[Trace] = []
         self.enable_monitoring: bool = enable_monitoring
         self.enable_evaluations: bool = enable_evaluations
-        self.class_identifiers: Dict[
-            str, str
-        ] = {}  # Dictionary to store class identifiers
+        self.class_identifiers: Dict[str, str] = {}
+
         self.span_id_to_previous_span_id: Dict[str, str | None] = {}
         self.trace_id_to_previous_trace: Dict[str, TraceClient | None] = {}
         self.current_span_id: Optional[str] = None
@@ -1498,7 +957,16 @@ class Tracer:
         # Initialize S3 storage if enabled
         self.use_s3 = use_s3
         if use_s3:
-            from judgeval.common.s3_storage import S3Storage
+            if not (
+                s3_bucket_name
+                and s3_aws_access_key_id
+                and s3_aws_secret_access_key
+                and s3_region_name
+            ):
+                raise ValueError(
+                    "S3 configuration requires bucket name, AWS access key ID, secret access key, and region name"
+                )
+            from judgeval.common.storage import S3Storage
 
             self.s3_storage = S3Storage(
                 bucket_name=s3_bucket_name,
@@ -1506,6 +974,7 @@ class Tracer:
                 aws_secret_access_key=s3_aws_secret_access_key,
                 region_name=s3_region_name,
             )
+
         self.offline_mode: bool = offline_mode
         self.deep_tracing: bool = deep_tracing
 
@@ -1613,7 +1082,7 @@ class Tracer:
         name: str,
         project_name: str | None = None,
         overwrite: bool = False,
-        rules: Optional[List[Rule]] = None,  # Added rules parameter
+        rules: Optional[List[Rule]] = None,
     ) -> Generator[TraceClient, None, None]:
         """Start a new trace context using a context manager"""
         trace_id = str(uuid.uuid4())
@@ -1634,23 +1103,19 @@ class Tracer:
             name,
             project_name=project,
             overwrite=overwrite,
-            rules=self.rules,  # Pass combined rules to the trace client
+            rules=self.rules,
             enable_monitoring=self.enable_monitoring,
             enable_evaluations=self.enable_evaluations,
             parent_trace_id=parent_trace_id,
             parent_name=parent_name,
         )
 
-        # Set the current trace in context variables
         token = self.set_current_trace(trace)
 
-        # Automatically create top-level span
         with trace.span(name or "unnamed_trace"):
             try:
-                # Save the trace to the database to handle Evaluations' trace_id referential integrity
                 yield trace
             finally:
-                # Reset the context variable
                 self.reset_current_trace(token)
 
     def identify(
@@ -3019,7 +2484,6 @@ class _TracedSyncStreamManagerWrapper(
         return self._original_manager.__exit__(exc_type, exc_val, exc_tb)
 
 
-# --- Helper function for instance-prefixed qual_name ---
 def get_instance_prefixed_name(instance, class_name, class_identifiers):
     """
     Returns the agent name (prefix) if the class and attribute are found in class_identifiers.
