+++ src/judgeval/tracer/__init__.py
@@ -1,3 +1,4 @@
-from judgeval.common.tracer import Tracer, wrap, TraceClient, TraceManagerClient
+from judgeval.common.tracer import Tracer, wrap, TraceClient
+from judgeval.common.tracer import TraceManagerClient
 
 __all__ = ["Tracer", "wrap", "TraceClient", "TraceManagerClient"]
